Fix several critical bugs and reformat based on AOSP code style.
Fix a bug in AppManager that causes a failure to install an app
in some API levels. Reset the minimum required Android build
version to 18. Change the tested app in AppManagerTesti since
since API 18 image has no email app. Fix a bug that fails the
scrolling in setting menu in API 18 and 19 (related: b/28622358).
Reformat everything based on AOSP code style which requires a tab
size of 4.

Change-Id: I35ab63db413f94bc1144538e69d84ce2af939152