[logparser] Flakiness_pipeline: Support clean runs.
This CL adds support for "seal"ing the pipeline, to indicate that
current processing is complete and the next attempt to move the pipeline
forward should clean the workspace and start again.
This CL also adds support for cleaning a sealed workspace.

Finally, this CL adds an example_step to demonstrate the working of the
pipeline.

After this CL, sequential runs of the flakiness_pipeline.py script will
result in the following changes to the workspace.steps table:
- Start from nothing
- Run 1: example_step --> started
- Run 2: example_step --> finished
- Run 3: seal [complete --> success]
- Run 4: Cleans up workspace and goes back to step 1, i.e.,
         example_step --> started

BUG=b.android.com/209673

Change-Id: I3f200ab746440a1b0b0ae5d5abd8203f6e049471