Allow emulator CTS tests to be excluded
Some CTS tests have been found to have persistent negative side-effects:
they may either crash the emulator, or leave it in a state of diminished
performance.  In order to get useful information from the rest of the
test run, this CL allows us to specify tests to exclude.

The set of tests to exclude is expressed as a data structure in the
test_cts.py file.  The mechanism we have available is that a CTS test
plan file specfies a set of test packages in the plan, but also allows
you to specify individual tests to be excluded.  The data structure is a
dictionary mappin plan names (e.g., "CTS") to another dictionary, this
one mapping test package names (e.g., "android.calllog") to a sequence
of excluded tests in that package (e.g.,
"android.calllog.cts.CallLogBackupTest#testSingleCallBackup").

When we're asked to execute a test plan, we check whether that plan has
exclusions in this data structure.  If so, then we create a new test
plan, derived from the .xml for the original.  We assume the original
plan has no exclusions.  We copy line by line; if the line specifies a
test package with exclusions, we insert those exclusions in the output.
We direct cts-tradefed to execute the newly-created test plan.

I've tested (via the test code now commented out in the file) that the
translation code produces the desired output: for the current CTS.xml,
and the exclusions given in the CL, it produces the desired output: a
CTS-emu-stable.xml file that differs from the original only by the added
exclusions.  If there are suggestions on how to test further, I would
welcome them.

Note: I also eliminated what seemed like a redundant second call to
create_test_case_for_avds().  If I'm missing a reason this might be
necessary, let me know.  Also, I'm not completely fluent in python, so
feel free to critique coding style decisions.

Change-Id: I57c0154439760fac4019adf5aeaa202031cb49e4