Update flakiness results based on tests run 2016-06-21_19-05-55
Until a few days ago, CTS tests were not running to completion.
These results are based on the few tests that were run after
that issue was resolved:

2016-06-19_10-46-23  build_0-rev_2988904
2016-06-20_17-31-08  build_3-rev_2991947
2016-06-21_19-05-57  build_6-rev_2997187

Summary of the information in this file:

    751  Bad
    127  Flaky
     63  gotBroken
 ------
    941  Total bad

    157  gotFixed
 119881  Good
 ------
 120979  Tests that were run

   6036  Not executed
 ------
 127015  Total tests

I have changed the format of this text file. The changes should make
it easier to compare revisions *in the future*. (But the changes make
it very difficult to compare with versions prior to this one.)

The format changes are:
  1) Put the flakinessResult on the same line as the fullName.
     This helps keep 'diff' tools sync'd.
  2) Alphabetize using sqlite3's sort order. (Rather than picking
     some arbitrary sort tool each time.)

Change-Id: Id3507787183c4d640e51f55c295fd92adb823a33