Allow CTS tests to be sharded N ways
This CL modifies the test infrastructure to allow the CTS tests to be
"sharded" -- not in the sense of running concurrently (though that could
be possible), but rather in the sense of running 1/Nth of the test
packages, then killing and restarting the emulator instance.  The hope
is that this makes the tests more robust, limiting persistent effects
that tests can have on subsequent tests.  Currently the sharding factor
is hardcoded at 10.

We take the base CTS.xml plan file as input, and generate N derived
CTS_i_of_n.xml plan files.  We execute each of these, and combine the
testResult.xml files they produce into a single top-level XML file for
the overall test run.

I also refactored a fair bit for testability, and wrote four tests to
try to ensure that I got this right, without actually running the new
scripts.

Testing: all the added tests pass.  In addition, I modified the script
to run a different plan file, with a different test sharding: it had 9
packages, sharded 3 ways.  It ran three packages in each test shard, and
seems to have correctly merged the output XML files into a single one
with all the results.

Change-Id: Ia401794303566d6984f944b511b9af576cf3de4a