Change the location of testing logs
The original location of testing logs is under the instrumentation
app's external storage dir, which will be wiped by gradle testing
tasks when the whole test suite is done and the instrumentation
app is uninstalled. So, in order to pull testing logs after gradle
testing tasks, we change the location to a place outside of app's
specific storage.

Change-Id: I8a1799a5c60275f92393f5dccb46a50989bf3325