Remove one excluded test; record bug for other.
This CL removes the exclusion of the test
  android.calllog.cts.CallLogBackupTest#testSingleCallBackup
from our CTS runs.  This was excluded because it triggered a kernel
panic.  It should have been restored a while ago; the underlying problem
has been fixed.

Also, adds a comment indicating the bug that goes along with the
remaining excluded tests.  We should always have such a bug when we
exclude a test.

Change-Id: I6df436fededbf8f961c90d307a685ffa10b4066f