Fix system image UI test by waiting for bookmark object to appear.
Test could not find the bookmark due to bookmark object not rendering
quickly enough on the buildbot. The test now waits up to 3 secs for
the bookmarks to load and then up to 3 more seconds for the new
bookmark to be found before veriyfing. The test also deletes the newly
created bookmark to clean up the test to prevent any false positives.

Test: refactor CL.
Bug: b/31991499
Change-Id: I2d3ed6f40c7290b0ecd84c1abecec3448bdbf579