Fix system image UI test for AssertionError: Failed to find the VPN lock icon after starting VPN.
The expandable notification tray was causing an issue in API 24. The VPN icon was not present in
the unexpanded view (or the expanded view, for that matter). For API 24, the test will now expand
the notification tray, and then search for the "Network may be monitored" text
(instead of the VPN icon) to complete the verification.

Test: fix test
Bug: b/30376641, b/31991033
Change-Id: Ic0b6166d5dbe1fb3f1a1fe326d08183ad1b3df91