Remove unused code in scripts/common/presubmit/constants.py
The file is folked from internal buildbot. But a bunch of code
is not used on external buildbot.

Because master_gen will be invoked by every slave machine which
leads to run constants and create a AgentLib based on local gitcookies,
the unused code removed in the CL will require every slave machine
to have .gitcookies locally, which is unncessary.

AgentLib is only created and used by recipe sysimage_release_psq,
which means only the slave assigned to run the test needs
to have this file locally.

Change-Id: I759cbb551341163d541dbb01353120178b25f132