Remove API 24 defaults images from UI and console tests config
Test configs must sync with existing pollers, or it could cause a
failure that a test fails to find image and create AVDs.

So far, we do not have any poller for API 24 defaults. So, we should
remove them from test config. We will add it back until the pollers
are ready.

Test: N/A
Change-Id: I318cb148c2d96077176d101846d7066651eae288