Updates code to properly handle remote fileIO on W
Windows machines have their local filepaths listed in the traditional
Windows format ('\' seperate).  Unfortunently, the code currently
does no processing of this path, and as a result we are attempting to
copy files to Unix hosts that are formed in this way.  This checkin
aims to resolve this issue by converting paths to a Unix suitable path
when talking to a remote host.

Change-Id: Ia1288d5abd9cce6fc05f6e618e52c9b2a47310eb