Workaround for Windows SCP behavior
Windows SCP has the 'interesting' property of not playing well with
filepaths between Windows and Unix.  For example, when I scp the folder
'foobar/foobet' to a *nix host, it actually creates a folder named
'/previous_path/foobar\\foobet'.  Which is silly.
This CL includes a quick fix for this (copy the folder directly, not
by addressing its path, and then reverting when done) until a better
solution can be discovered.

Change-Id: I3aa765ff335b22426f7c7287fd3a3805e31b1cd2