Updates launched of AVD to occur in main thread
Previously, a seperate thread was spun up where the AVD instance was
run.  This often caused a timeout to occur during the testing, as the
main thread would be unable to poll the child thread within the needed
time frame.  This CL changes this behavior, now only starting up a new
thread for the logcat process, running the actual AVD within the main
thread.  This is consistent with the way CTS is currently tested.

This is based upon a CL done by @bohu.

Change-Id: Id6b8b7682d04ce141b61fda35e252f64eaa18a26