Print the progress of an adb stress test
An adb stress test runs a large amount of iterations, and there could
be no print out for hours (depending on the duration set) if there
is no error. However, buildbot uses stdout/stderr polling to decide if
a step is still active and will kill it if it is inactive for one hour.

To avoid it, we call adb stress test with --progress option so that
after every iteration, a progress status will be printed out to
make sure that buildbot can poll something from stdout within one hour.

Test: recipe_simulation_test.py train
Change-Id: I7994bb60a9e95dcc2038c1624b2eb992a9ec26fc