Fixes scp call in Unix after Windows fix
This CL changes a error in the Unix codepath introduced in CL307887.
We need to be in the directory of the existing file when scp'ing the
results due to windows SCP implementation.  This CL removes the
is_windows check from this logic to fix a issue with code flow seen on
Linux machines.

Change-Id: I1a400dcf1d9f7aff3d7c9c4d504835d9e2a6e21f