Updates Buildbot to include CL list as step
This change is based upon reverted change 316898.
In that change, the CL list generation was added as a step.
However, in the event that prev_build was not properly passed
in the properties of the build, a uncaught exception would occur,
breaking the build.  This change adds a except to catch this exception.
This allows the code to be on production in the event the props json is
still improperly formed.

Change-Id: I84f283d19e6e4054c02261d556cbd33e989c8a2b