Updates ADT Builders to create a CL list for build
This update includes a new recipe step in the adt builder that builds
a list of all CL's that have been put into the build since the last
build for that branch.  This is done through calling into the
AndroidBuild RESTFUL api.

API relies on a new file, .ab_credentials.json, which will be manually
places on each slave.

Change-Id: Id9e553011abefa273989da93d9fe2d22feb0349e