Disable attempting to send email through SMTP
Email is not working on these machines due to route issues.
Trying to send this email is costing us time on the master while
it attempts to send the message.  I have modified the code to not
attempt to send any message, and to print out a message saying as such.

Change-Id: I6bea35ec054f7cea648fa82c10e56ef223788d15