Fix flaky time zone test.
Fixed the issue where if automation clicked the "Select time zone" button before
it becomes enabled, the click would not have an affect. The testw ould proceed thinking it's
on the next screen and would cause the test to fail.

Added code to ensure that the test wouldn't try to click on "Select time zone" until the button
was confirmed to be enabled. This will prevent the timing issues seen when the test fails.

Test: manual
Bug: b/34814446
Change-Id: I0c1ba3716c17a5cc26228c7a417b3cf073a12c68