Fix bookmarkWebSiteInBrowser test.
Test only fails on the bots so this fix is a guess based on a screen shot of what might be happening.
Test on the bot seems to indicate the UI freezes on the Chrome browser terms and services screen.
The fix extends the run time of the test, then adds a 3 sec wait in several locations to account for any objects
that's not loading quick enough. Also dismisses a request to sync the bookmarks. API24G uses the chrome browser
instead of the Android web view.

Test: none
Bug: b/35365619
Change-Id: I341b263da0acefa72eaa74eee59ea6606eb4e64c