Fixes issue with Windows systems carriage return
Windows has a carriage return in its files, so when adb does its
device connected test, windows was seeing an extra 'device' (which
was just a badly split line).  We now test for the platform of windows,
and if we detect it properly parse out '\r\n' instead of just '\n'.

Change-Id: Ie653a63b8067b197d42cb399d96eaf9c5a20734b