Adds support for Swarming based Emu Boot Test
Adds neccessary files for a Swarming-based Emulator Boot Tester.
When a new emulator build is built by TreeHugger as part of PSQ, the
Swarming server will send a request to
emu_test/psq_boot_test/psq_boot.py to initiate a PSQ boot test.  This
result will be returned to the swarming server and intregrated as part
of the basic PSQ infrastructure.

Change-Id: Ie706df599ee0282ebd7341f46fdcc04b7af3c819