No longer fail when gs_multi_poller fails to find
Currently, when gs_multi_poller fails to find one of the expected
gs://paths, it will lock into a cycle of failure.  It will detect that
the buildId is invalid, and return.  The issue is that the buildId will
never incremenet.  It will always keep checking that failed buildId,
effectively locking down the poller.

This change changes the behavior to no longer return a empty list of
objects when a platform build number is missing.  We simply log this to
a file, and continue with what is currently present.

Bug: b/67591163
Change-Id: Idf7dc86939240cfab5fad34283bc17a479031650