Change the ADB stress test framework to parallelize the test suite.
This allows each device to run its iteration independent of other
devices. Currently, each iteration is run on each device in lock-step.
One iteration of the test is run in parallel but the next iteration does
not start until after each device has completed the previous iteration.

This change allows each device to begin the next iteration as soon as it
is done with the previous one. Faster devices can run more iterations
than slower devices in the given time.

BUG: b/70404604
TEST: local
Change-Id: I5acb9bc220f79182ea87f94a2bf7fff514e8cf0c