Fix bug where AppLauncer.launch() cannot launch off screen apps.
UIScrollable.scrollIntoView() not able to verticle scroll the UI when
the app icon is off of the UI screen. Instead the keyboard popsup. This
CL fixes the issue by using the fling gesture to fling the screen
forward by a max of 5 swipes.

The bug happens on any phone in which the Settings icon is not visible
when the App Menu is loaded. The issue happens on a Pixel due to the
narrow width of the screen.

The bug is a scrolling bug that seems to be only occuring in API 26
because AppLauncher.launch() cannot launch the Settings app when the
Settings icon is offscreen due to UIScrollable.scrollIntoView() failing
to scroll to the target selector.

TEST: local
BUG: b/72566802
Change-Id: If6862ebc40166f96ada5f3ac83d9f61a3f29e1ce