Add UI test for video recording and fix camera test.
Add a UI test to verify the video capture functionality of the camera
application.
Fix a camera test bug. The original test assumes the camera folder
exists prior to the test running, but the camera folder is created only
after the first photo is taken via the camera app.
Tests support APIs 24+

TEST: local
BUG: b/74022221, b/72697181
Change-Id: I397579d5efcb3242c1fa07ab73c71e466b4c1e02