Fix UI test confirmNowCardsPageOpen.
There is no internet connection in airplane mode, thus the now cards
page does not connect. The fix checks to see if airplane mode is on, if
it is, the test disables airplane mode prior to running the test.

The ideal fix would be to write a utility to testf or internet
connectivity.

TEST: local
BUG: b/30819490
Change-Id: I0c99c5c97ea0d18f72ffb6a4691fc06ac08cfefe