Create common util to install demos app and improve error handling.
Create a common utility ApiDemosInstaller and improve error handling for
three test methods in SettingsTest and one test method in APIDemosTest.
Those tests will only proceed if the isApiDemosInstalled flag is set to
true, if the condition is not met there those tests will be skipped.
The util searches for 'com.example.android.apis' in the list of
installed packages on the image.

TEST: local
BUG: none
Change-Id: I11552d2a165a4a489703d3a2ed18977bb97bf09e