Updates dotest.py arguments to be more specific
Instead of single letters and smaller words, use proper flag names for
the different parts of dotest.py.  This makes it easier for a user to
see what dotest.py is actually doing, as instead of seeing '-p
"test_.java"' for instance, they will see '--pattern "test_*.java"'
which is more readily identifiable.  Update appropriate call chains for
this change.

Test:  Local simulations.
Bug: 77235882
Change-Id: I58001bbf6263f5adfee9d24c9ab8194e06b0ecce