Update rmtree on WindowsError to try permission ch
When we attempt to remove a existing AVD on Windows machines we are
encountering errors.  The errors appear to be due to some files being
marked read-only within the AVD. On windows this causes the
shutil.rmtree command to fail.  This checkin adds a new onerror call to
the function to attempt to set the file read/write and retry the remove.

Test:  Tested locally.
Bug: N/A
Change-Id: I3d5c351b2813d83c50912a6d1a782432e504a68f