Remove per-branch emulator selection
For some branches we appear to want to use a tot emulator image
(presumably for compatibility reasons).  However, the current
implementation does not properly perform this.  The current
implementation makes a assumption that a tot emulator exists within the
installed SDK.  This image is not automatically placed their by the
recipe, nor is it ever downloaded (presumably we could be performing a
boot test with a 1+ year old 'tot' image).  We must rethink this logic
if it is still required.

Test:  Ran Locally.
Change-Id: I7c70d2189dc8bcc7764b56cc76e994f52bd6fa29