Fix freezing issue when the emulator launches.
The freezing issue was caused by the stdin/stdout for the emulator
process being redirected to the streams, but those streams were never
read. The buffer was filling and then the emulator process was blocking
trying to write to stderr/stdout.

The code has been changed to launch the emulator process to not do the
redirect. Also the exception being thrown when the emulator boots is
fixed. The test now luanches the emulator, runs successfully and
terminates the emulator.

Test: local
Bug: b/111444824
Change-Id: Ic1684723c1302428d3ddad7732b2fecaa8cb2fe7