Fix error for launch AVD test regarding invalid case in filter_dict.
Invalid case error in filter_dict flag when gpu=yes.
The create_test_case makes a recursive call to itself if gpu=yes in the
avd config. The same avd config is passed to the recursive call, except
that the value for gpu ("yes") is replaced with swiftshader_indirect.
This was causing the valid_case function to complain with the given
filter since the filert's value of "yes" for gpu no longer matched the
value for gpu in avd_config.

test: local
bug: b/111444824
Change-Id: I738ef327d3e5d2970401e0d294cc867777a4f353