Update slaves to reference new IP addresses
The ChromeOS lab went through a full move over the weekend.  Because of
this move, we need to reboot and reset all go/adt-builder machines.  The
move assigns new IP Addresses and hostnames to every machine, however
due to a lack of DNS in the new lab we can no longer reference machines
by hostname.  So we connect to every machine via IP address.

Test:  Ran local.
Change-Id: I14f9c1a061864c4e490d4f0530c0b7556fc9789f