Fix for test script
Boot test for API 21 is failing because API is not mentioned in the
config file.

Instead of running for all 3 GPU configuration for all the lines in
config, now the tests are more controlled by config files. If different
GPU configs needs to be tested, then they have to be specified in config
file.

Test: Local
Change-Id: Ifab5629daa1d7b5621ca180d3ebbbb869c63cd06