Set title for command window
For windows, set title for command window so that it can be killed
easily using taskkill in case the test is hanged. For the UI test,
if it takes more then 2 hours, it will be killed. Similarly, for the
boot tests, if it take more then 1 hour to complete, it will be killed.

Test: Local
Change-Id: I32d62afc75d6a89b766bb2f77ba1c8afd4b27ec7