Wait for object rendering before clicking on objects.
SettingsTest.setTimeZone on GCE takes over 5 seconds to render after
launching the AVD. The first action is to click on the TV Launcher,
which is wrapped in a 5 sec wait. Increasing the time to 10 seconds to
allow the image to complete the initial render.

testBrowserLoadsSite is having issues clicking on the Accept and
Continue button on GCE. When Chrome launches, the button should be
clicked immediately. Wrapping the object in a 5 sec wait to see if the
issue resolves.

Test: local
Change-Id: I83d7560c97b164cee200fd03c5dab75d31ba28b3