Add a pre-test util to log into the play store prior to tests run.
This change groups all failing PlayStore UI tests into one failure as the failures tend to be for the same reason, not being able to login due to a UI change in the login process that is being handled outside of Android. Before any of the tests in PlayStoreTest runs, the test runner will check if it's a Play Store image. If not, then all the tests will be skipped to prevent failures on non Google API images. If the PlayStore is present, the framework will run the pre-test of running through the required steps of logging in a user into the PlaySTore. The result of the login attempt will be captured. The runner asserts on the PlayStore login, rather than during the execution of each test. No other tests in PlayStore will be executed and the normal assert error shutdown procedure will occur, which will log the error, capture a screenshot and dump the hierarchy within go/ab. This output should help locate where the test failure occured.
This is used to smoke out root cause of the test failures.

The PlayStore login logic has been removed from the tests themselves to reduce the logic, run time and potential failure points for each individual test. The CL requires modifications to the common PlayStoreUtil.loginGooglePlay utility that all PlayStore APIs use. The @before will be executed prior to the tests for each API level.

Test: local
Bug: b/123295827
Change-Id: I5ce12e05ba19e3b11350808a119c5f2f69d7cea9