Reduce wait time to prevent time out.
When UI objects aren't found, the code will wait for the UI to render
before moving forward. If those waits are too short, the test will fail
with a UiObjectNotFound exception. However, if the waits are
collectively too long, the entire test can time out. Since there were
several UO changes in API 28 Rev 8, several waits were triggered. This
change reduces the length of a wait constant that was repeatedly used in
the Vpn test on the Mac.

Test: local
Change-Id: Id375fec5e41bc0620bed5f84f661f529479a4acd