Refactor Map UI test to use key event instead of UiScrollable.
The UiScrollable could no be found in latest API 28 UI change, which
caused the test to fail. Refactored the test so it no longer uses the
sroll view. Instead the test proceeds to the same end point by using a
simulated key event. This makes the test shorter and faster.

Test: local
Change-Id: Ieb16486766af63b64a8e263ab0a4c0e7cbc0966a