Fix UI test to bookmark web site in browser in API 27, 28 and 29.
AppTest.bookmarkWebSiteInBrowser was failing since the Chrome bookmark
could not be deleted. This edit changed the delete process in API 28 by
clicking on the trash can in the final step. Removed the non-Chrome
generic browser code path as this is no longer valid in API 28. Applied
same changes in API 27 and 29.

Test: local
Change-Id: I656259ad0fb82f5f535f65a30a9a439e45515dc8