Fix sign in error.
Account sign in page isn't properly determining if the user needs to
sign in. The fix extends the wait for the sign in button to allow the
page to load after the spinner. A check was added for the create account
string as a fallback to determine if logging in is needed.

TEST: local
Change-Id: I8af33dff07622ec8ac0e0cdf4bb75b20096a21b8