Fix ApiDemosTest, AppTest, CameraTest, SettingsTest, VpnTest
Fix tests for API 29
ApiDemosTest - the apk is preinstalled before starting the tests.
AppTest - the apk is preinstalled before starting the tests.
CameraTest - the apk is preinstalled before starting the tests.
SettingsTest -
VpnTest - the apk is preinstalled before starting the tests.
GoogleAppUtil - fixed work flow for logout by changing to using regex

Delete .idea files whicha re generated bt the ide and not supposed to
sit in the repo.

Bug: 138744604, 138401038, 138386930, 139482815
Test: Local
Change-Id: Iff24e819b70d89cd45ff9614f8d27a506b90ea14