Extra screenshot for debugging
Many times, we see "in transition" screenshot when a test fails. For
debugging purpose, now we are taking another set of hierarchy dump and
screenshot after a wait to figure out if a longer timeout would help or
not. This will also show us if the emulator was able to move past the
click action that it was performing at that time.

Test: Local
Change-Id: Ib005e86492094937b81e1105f256fe5e2e13c84c