Integration test for icebox
This is the integration test for icebox. It boots the emulator, install
an APK and its failing test, run the test with debug, attach icebox
debugger, check if test finishes and fails, and check if icebox takes
the test failure snapshot correctly.

It needs some wait between starting the test and starting icebox.

Change-Id: Ic7a94b42426bd3ecd7e0d8e69d260d82fb77bf5b