Add support for E2E testing of gRPC endpoint
This adds a tox/pytest based for e2e grpc tests.
These test closely resemble how android studio will interact with the
emulator, that is.

- It launches the emulator in the same way as studio would.
- It uses the same discovery mechanism to find the emulator.

You can run the tests as follows:

make

  py2: commands succeeded
  congratulations :)

The README.md contains detailed instructions on how to run tests
standalone
For now both Python2 (needed for the build bots) and 3 are supported.

It contains a set of starting tests and a few that are disabled due to
them exposing real issues.

Change-Id: I68e723f2dc5194bf1cd492504442240595fe064a