Add performance tests for gRPC vs. Telnet
This adds a performance testing framework that can be used to measure
how fast the gRPC endpoint is v.s the telnet endpoint.

It has the following series of tests:

- test_mouse_perf_host_host_.* Tests to see how long it takes from the
sending of an event to arrive at the emulator. This only measures the
arrival at the agent and not the dispatching to the guest.

[Telnet|Grpc] --> Emulator Agent

This compares the actual time stamp of send against actual timestamp of
the console agent.

- test_mouse_perf_host_guest.* Tests to see how long it takes from the
sending of an event to arrive in the guest. This measures the complete
traversal time to arrival in the kernel. Unfortunately due to clock skew
between guest and host these numbers are completely useless.

[Telnet|Grpc] --> Emulator Agent --> Guest

This compares the actual time stamp of send against actual timestamp
reported by getevent -t inside the guest.

- test_mouse_perf_wall_.* Tests to see how long it takes from the
sending of an event until the detection of an event.

[Telnet|Grpc] --> Emulator Agent --> Guest --> event observed in ADB

You can run the individual tests by:

pytest -k "test_mouse_perf_host_host_"

For example:

----------------------------------------------------------------------------------------- benchmark 'mouse-host-host': 2 tests ----------------------------------------------------------------------------------------
Name (time in us)                           Min                   Max                  Mean              StdDev                Median                 IQR            Outliers         OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_mouse_perf_host_host_telnet       375.9861 (1.0)        881.1951 (1.0)        487.0951 (1.0)      105.8949 (1.0)        473.1417 (1.0)       97.2748 (1.0)           5;3  2,052.9871 (1.0)          40           1
test_mouse_perf_host_host_grpc       1,374.0063 (3.65)     2,092.1230 (2.37)     1,705.1041 (3.50)     170.6849 (1.61)     1,714.8256 (3.62)     204.0863 (2.10)         11;0    586.4745 (0.29)         40           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

Bug: b/161233750
Change-Id: I59e8d0147304f8fac5fcfd57b355bc94187b7cda