Fix all broken SettingsTest in API 28.
Update all tests to use the correct version of the APIDemosInstaller.
Refactored testLocationSettingsPage to use the Application Launcher to launch the target page.
Fix testPhonePermissions by using the correct selectors to get to the phone screen due to change in resource id. Expanded the test to include an attempt to place a phone call.
Updated testMapsPermissions to use the correct version of the utility method.
Fix tetCameraAppDisabled to use the proper argument signature.

Add ignore annotation to filesDeleted due to updated security measures in API 28, which prevents the test from deleting system files.

Change-Id: Idbaa2a273110f577f75197d175bdd9a118d46ae1