Enable embedded emulator tests in presubmit.
This enables the embdded emulator e2e as part of the build and disables
them from the test targets. Builds will now fail if the presubmit tests
do not succeed.

The tests will only run on linux and take around +/- 7 mins to complete.

Tests that are marked as a performance test will not be run.
Some tests are marked as disabled as the required APIs are disabled.
Some tests are makred as disabled, since they exposed some real
breakages.
Some tests are marked as disabled as they rely on adb to measure
success, these tests are too flaky.

Includes a set of fixes to make sure that all the tests are green.

Bug: b/178497994
Change-Id: I36a3e960c5c6f7045ce8d3d7696570cb9f817087