Fix testPhotoCapture and testVideoCapture in API 30
Update camera utilities in API 30 to fix testPhotoCapture and testVideoCapture. Both tests failed during the file deletion process. Updated the names of the file folders and created a new utility to delete the test file. The method identifies camera files to be deleted by their prefix and uses a long press to select them.

Change-Id: I7c4b4d86b82f05c920b14ff9b175e523fcee6c0b