Vpntest and Network test for api 30
Provide alternate device connected text to fix VPN Test on API 30.  The verification string was changed in API 30.
Added a new test and removed disabled3GData since the test assumes 3G mode is the Preferred Network type. In API 30, 5G is the preferred Network type. Rewrote test to be more relevant to API 30.

Change-Id: I50890a99db5153a42d2af03bcc35ff36cdf6bd06