[gRPC] e2e test for rotation
This adds two additional end 2 end tests that validate the rotation
behavior:

- The animation app now listens for rotation and reports the observed
rotation. The test verifies that the reported rotation coincides with
the rotation we set.

- The animation app now draws a square with the color #DECEBE in first
quadrant. The test validatates that the square ends up in the proper
quadrant when the device is rotated. This makes sure that the rendering
engine is applying the correct translation.

Note: The errors were not caught due to us disabling e2e tests as part
of presubmit due to stability issues b/181834805.

Bug: b/182198601
Change-Id: I414d7430882192dd7d60cb158595a881ba0d3103