[gRPC] Improve stability of e2e tests.
Previously we relied on the @pytest timeout markers to detect failures
in our streaming grpc calls. This works well in the sense that we detect
hanging calls, but it forcefully terminates them.

We now change the api slightly to explicitly set a timeout on how long
we wish a call to last, and turn the StreamingCall into an actual
iterator. This has 2 benefits:

- We no longer have the clunky queue to iterator modifcation
- We actually timeout on the gRPC calls, enabling us to use the flaky
module to attempt re-runs for our flaky tests.

Change-Id: Ie0f38668e7c81d737d96cc4ae28b04482cd18f21