Fix Network IO Test to support API 32 due to new UI changes.
API 32 consisted of several timing based errors when running the Network
IO UI test. EnableAirplaneMode was failing at the end of the test on
clickAndWaitForNewWindow. There was no new window to end the wait as
this was the last line of the test so the test hung at the end of
execution. Fixed the test by sending the test home.

There is an intermittent crash in the emulator when running
toggleCellularDataMode.

Change-Id: Ie918832d4e69fd1cb179da76466b4e40ac2001b0