Improve logging and stability
This improves the logging, documentation and fixes a small series of
stability issues.

We now properly shutdown the emulator when the emulator object goes out
of scope.

We clean up the launcher script to not log all the sdkmanager licesnes
and will now log all the pip resolver issues that might happen.

We produce a test.xml that is compliant with what sponge expects,
resulting in less storage and more details in sponge itself for our
failures.

Change-Id: Iae8788ae48f8b78ef9a9848f0ee8632c9244d2d5