Fix the package installation utility that broke after Android 11.
isPackageInstalled returns inaccurate results for newer APIs since it
uses PackageManager.getInstalledApplications, which does not work for
Android 11+. This fix will install the APK regardless. If the APK is
already installed, it will be treated as an update. The routine will be
used without having to know the app installation state in advance.

BUG: b/250777894
Change-Id: Ie0955cb3e11ac66fc99925b36a70dd57387ee88e