Enable working with .gz images
This enables the tests to work with compressed images that are checked
into the system_image branch.

This significantly reduces the run time of the test as we do not have to
download system images using sdkmanager.

The configure script will now set the ANDROID_SDK_ROOT to the internal
one, making sure you will use the proper sdk root during development.

Bug: 247856786
Change-Id: I034ac926202087c2b3d1af42f4cc1b0a8095d0cc