Fix the rotation tests
This fixes the rotation tests by looking for the proper pixels.

We now transform the incoming pixels to a pillow image that we inspect
for colors and pixels.

We also turn the rotation test into a parameterized test, making it
easier to find issues if they are related to a particular angle.

Bug: 246780175
Change-Id: I5189d7864ceb28690e486ae3d9be177b67790f2d