Fix windows psutil import issue
The windows python interpreter in AOSP ships with Python310.dll vs the
expected Python.dll. This breaks the windows tests that rely on the
psutil package. This packages needs Python.dll, which does not exist.

We now create a symlink between Python310.dll <-> Python.dll to make
sure we can load the psutil module.

Bug: 265843618
Change-Id: Id96971e77c6d73761c2be43cf8c017ec54889cbe