Do not set defaults for --emulator flag
The defaults that were derived relied on ANDROID_SDK_ROOT or
ANDROID_HOME, which are not always set to values that actually contain
an emulator executable, resulting in failures when the parameters is not
present.

We now derive a default if the parameters is not set during the launch,
and will default to the "common" build directory.

Bug: 266629617
Change-Id: I9fee0965cbfbf3e0fc41c7fd7c340a5efa3e0347