Enable verbose logging on darwin
We are seeing failed pip install on darwin-x86 and are running into
timeout issues.

We now set the default logger to verbose under darwin, which should give
us some insight into where we are getting stuck.

We also update the logformatter to provide more detailed information and
the binplace directory of our xml results.

Bug: 261042155
Change-Id: I86fe6432e2b11f5d1dc134e8fb0333033208bc12