Enable usage of custom python interpreter
The AOSP python interpreter is crippled and does not have symbols.
This makes it difficult to make use of additional tools to inspect
potential failures.

We now add the --no-aosp flag to use the python interpreter found on
the path.

Includes a set of run and configure scripts for windows.

Bug: 253078934
Change-Id: I269fb7368a22df6a8a7aa8cb6833c49aad39cd6f