Stabilize sensor tests
The sensor api is eventually consistent. Meaning that setting the sensor
value happens on a different thread than the one used to set it. This
makes it possible to have a sensor read while it has not yet been
propagated.

We now update the test to poll the sensor value a few times before
declaring a failure.

Bug: 269137594
Change-Id: Iaf08b81089884816106c04b3630ba025edb804bc