Use renamed symbol file
We renamed the symbol artifact. This should make it available on the
build bot.

We now log which files we are considering and use a regular expression
to pick the proper file.

We also weaken the asserts in the crash analysis as some compilers are
optimizing better than others, resulting in slightly different
callstacks.

Bug: 271168786
Change-Id: I276d121efa76cf52a8d6a82ba7cdf226540b0e0d