Use virtualenv in Windows
This add support for using the "virtualenv" module in Windows. This
module provides a superset of the venv packages that was removed from
the windows release.

In order for this to work we provide a "ssl.py" stub that contains no
functionality. This enables us to use virtual environments in windows.

Bug: 274000767
Change-Id: I6aafe77a5b565fbe7329be542083f7852b1423e5